requirements.txt
pbr>=0.5.21,<1.0
# Horizon Core Requirements
Django>=1.4,<1.6
django_compressor==1.3
django_openstack_auth>1,<=1.1.9
eventlet>=0.13.0
kombu>=2.4.8
lesscpy>=0.9j
iso8601>=0.1.8
netaddr
python-cinderclient>=1.0.6
python-glanceclient>=0.9.0
python-heatclient>=0.2.3
python-keystoneclient>=0.3.2
python-novaclient>=2.15.0
python-neutronclient>=2.3.0,<3
python-swiftclient>=1.5
python-ceilometerclient>=1.0.5
python-troveclient<1
pytz>=2010h
oslo.config==1.7.0
# Horizon Utility Requirements
# for SECURE_KEY generation
lockfile>=0.8
pycrypto==2.6.1
